chore: 루티 1.0.0 배포 (#689)

+++ backend/spring-routie/src/main/java/routie/place/service/PlaceService.java
@@ -10,8 +10,11 @@ import routie.place.controller.dto.response.PlaceCreateResponse;
 import routie.place.controller.dto.response.PlaceListResponse;
 import routie.place.controller.dto.response.PlaceReadResponse;
 import routie.place.domain.Place;
+import routie.exception.BusinessException;
+import routie.exception.ErrorCode;
 import routie.place.repository.PlaceClosedDayOfWeekRepository;
 import routie.place.repository.PlaceRepository;
+import routie.routie.repository.RoutiePlaceRepository;
 import routie.routiespace.domain.RoutieSpace;
 import routie.routiespace.repository.RoutieSpaceRepository;
 
@@ -23,6 +26,7 @@ public class PlaceService {
     private final PlaceRepository placeRepository;
     private final RoutieSpaceRepository routieSpaceRepository;
     private final PlaceClosedDayOfWeekRepository placeClosedDayOfWeekRepository;
+    private final RoutiePlaceRepository routiePlaceRepository;
 
     public PlaceReadResponse getPlace(final String routieSpaceIdentifier, final long placeId) {
         final RoutieSpace routieSpace = getRoutieSpaceByIdentifier(routieSpaceIdentifier);
@@ -36,7 +40,7 @@ public class PlaceService {
             final PlaceCreateRequest placeCreateRequest
     ) {
         RoutieSpace routieSpace = routieSpaceRepository.findByIdentifier(routieSpaceIdentifier)
-                .orElseThrow(() -> new IllegalArgumentException("해당 루티 스페이스를 찾을 수 없습니다."));
+                .orElseThrow(() -> new BusinessException(ErrorCode.ROUTIE_SPACE_NOT_FOUND));
 
         Place place = Place.create(
                 placeCreateRequest.name(),
@@ -57,7 +61,7 @@ public class PlaceService {
 
     public PlaceListResponse readPlaces(final String routieSpaceIdentifier) {
         RoutieSpace routieSpace = routieSpaceRepository.findByIdentifier(routieSpaceIdentifier)
-                .orElseThrow(() -> new IllegalArgumentException("해당 루티 스페이스를 찾을 수 없습니다."));
+                .orElseThrow(() -> new BusinessException(ErrorCode.ROUTIE_SPACE_NOT_FOUND));
         final List<Place> places = routieSpace.getPlaces();
         return PlaceListResponse.from(places);
     }
@@ -70,9 +74,6 @@ public class PlaceService {
     ) {
         final RoutieSpace routieSpace = getRoutieSpaceByIdentifier(routieSpaceIdentifier);
         final Place place = getPlaceByIdAndRoutieSpace(placeId, routieSpace);
-
-        place.getPlaceClosedDayOfWeeks()
-                .forEach(closedDayOfWeek -> placeClosedDayOfWeekRepository.deleteById(closedDayOfWeek.getId()));
 
         place.modify(
                 placeUpdateRequest.stayDurationMinutes(),
@@ -88,18 +89,21 @@ public class PlaceService {
     public void removePlace(final String routieSpaceIdentifier, final long placeId) {
         RoutieSpace routieSpace = getRoutieSpaceByIdentifier(routieSpaceIdentifier);
         if (!placeRepository.existsByIdAndRoutieSpace(placeId, routieSpace)) {
-            throw new IllegalArgumentException("해당 장소를 찾을 수 없습니다.");
+            throw new BusinessException(ErrorCode.PLACE_NOT_FOUND);
+        }
+        if (routiePlaceRepository.existsRoutiePlaceByPlaceId(placeId)) {
+            throw new BusinessException(ErrorCode.ROUTIE_PLACE_EXIST);
         }
         placeRepository.deleteById(placeId);
     }
 
     private RoutieSpace getRoutieSpaceByIdentifier(final String routieSpaceIdentifier) {
         return routieSpaceRepository.findByIdentifier(routieSpaceIdentifier)
-                .orElseThrow(() -> new IllegalArgumentException("해당 루티 스페이스를 찾을 수 없습니다."));
+                .orElseThrow(() -> new BusinessException(ErrorCode.ROUTIE_SPACE_NOT_FOUND));
     }
 
     public Place getPlaceByIdAndRoutieSpace(final long placeId, final RoutieSpace routieSpace) {
         return placeRepository.findByIdAndRoutieSpace(placeId, routieSpace)
-                .orElseThrow(() -> new IllegalArgumentException("해당 장소를 찾을 수 없습니다."));
+                .orElseThrow(() -> new BusinessException(ErrorCode.PLACE_NOT_FOUND));
     }
 }
